// src/generated/models/Pm_workflowchecklistconfigurationsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Pm_workflowchecklistconfigurationsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Pm_workflowchecklistconfigurationsstatecode = keyof typeof Pm_workflowchecklistconfigurationsstatecode;
export const Pm_workflowchecklistconfigurationsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Pm_workflowchecklistconfigurationsstatuscode = keyof typeof Pm_workflowchecklistconfigurationsstatuscode;

export interface Pm_workflowchecklistconfigurationsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  pm_isrequired?: boolean;
  pm_itemname?: string;
  pm_name: string;
  pm_workflowchecklistconfigurationid: string;
  "pm_WorkflowStepTemplate@odata.bind"?: string;
  statecode: Pm_workflowchecklistconfigurationsstatecode;
  statuscode?: Pm_workflowchecklistconfigurationsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Pm_workflowchecklistconfigurations extends Pm_workflowchecklistconfigurationsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pm_isrequiredname?: string;
  pm_workflowsteptemplatename?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  pm_workflowsteptemplate?: object;
  _pm_workflowsteptemplate_value?: string;
}