// src/generated/models/Pm_checklistresponsesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Pm_checklistresponsesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Pm_checklistresponsesstatecode = keyof typeof Pm_checklistresponsesstatecode;
export const Pm_checklistresponsesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Pm_checklistresponsesstatuscode = keyof typeof Pm_checklistresponsesstatuscode;

export interface Pm_checklistresponsesBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  "pm_ChecklistConfiguration@odata.bind"?: string;
  pm_checklistitem?: string;
  pm_checklistresponseid: string;
  pm_name: string;
  pm_responseflag?: boolean;
  "pm_WorkflowApprovalStep@odata.bind"?: string;
  statecode: Pm_checklistresponsesstatecode;
  statuscode?: Pm_checklistresponsesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Pm_checklistresponses extends Pm_checklistresponsesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  pm_checklistconfigurationname?: string;
  pm_responseflagname?: string;
  pm_workflowapprovalstepname?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  pm_checklistconfiguration?: object;
  _pm_checklistconfiguration_value?: string;
  pm_workflowapprovalstep?: object;
  _pm_workflowapprovalstep_value?: string;
}